[CELEBORN-20][REFACTOR] Extract CommitManager from LifecycleManager (#1050)
